Separate user related operations from authController
