--- src/main/java/com/inventory/server/model/user_permission/UserPermissionId.java
@@ -1,46 +1,0 @@
-package com.inventory.server.model.user_permission;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Embeddable;
-
-import java.io.Serializable;
-import java.util.Objects;
-
-@Embeddable
-public class UserPermissionId implements Serializable {
-
-    @Column(name = "id_user")
-    private Long idUser;
-
-    @Column(name = "id_permission")
-    private Long idPermission;
-
-    public Long getIdUser() {
-        return idUser;
-    }
-
-    public void setIdUser(Long idUser) {
-        this.idUser = idUser;
-    }
-
-    public Long getIdPermission() {
-        return idPermission;
-    }
-
-    public void setIdPermission(Long idPermission) {
-        this.idPermission = idPermission;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        UserPermissionId that = (UserPermissionId) o;
-        return Objects.equals(idUser, that.idUser) && Objects.equals(idPermission, that.idPermission);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(idUser, idPermission);
-    }
-}
