+++ src/main/resources/application.properties
@@ -7,11 +7,6 @@ spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.open-in-view=false
 
-# TOKEN CONFIGS
-# Default secret after colon, in case TOKENSECRET is not configured in Intellij
-# Will need to change on prod env
-api.security.token.secret=${TOKENSECRET:321654}
-
 # MULTIPARTFILE CONFIGS
 spring.servlet.multipart.max-file-size=1MB
 spring.servlet.multipart.max-request-size=1MB
